Moving CustomOutPlugin to out_directsound
